+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Translations: [简体中文](CHANGELOG_zh.md)
 
+# new
+
+decode:
+
+* improve: Improve to determine whether regional decoding is supported based on mimeType. Non-image
+  types directly return false. On non-Android platforms, based on the skiko version, it is
+  determined whether heic, heif, and avi types are supported.
+
 ## 1.1.0-rc01
 
 * change: SubsamplingState.disabledTileBitmapCache rename to disabledTileImageCache